Use alias instead of type declaration
This avoids the hiding of the underlying type and its methods. A
use case would be to access the methods of `soap.XSDDateTime`.
Consider the following code:

```
type StartDate soap.XSDDateTime

type FooBar struct {
    Date *StartDate
}

...

fooBar := FooBar{}
```

It is not possible to access `foobar.Date.ToGoTime()`. This commit
declares the `StartDate` type as follows:

```
type StartDate = soap.XSDDateTime
```
Which allows the complete access to the underlying type.